feat: Moved `TabsController` as a creator function
