+++ packages/plasma-core/src/components/Tabs/index.ts
@@ -1,3 +1,8 @@
+export { createTabsController } from './createTabsController';
+export type { TabsControllerProps } from './createTabsController';
+
+export { TabsContext } from './TabsContext';
+
 export { Tabs } from './Tabs';
 export type { TabsProps } from './Tabs';
